test: split platform-version and lite-SDK checks into their own files
The two guards added for #112 lived inside test/read-only.test.mjs, which otherwise only runs tutorials as subprocesses. Neither guard is a tutorial run, so they are moved to test/platform-version.test.mjs and test/lite-sdk-versions.test.mjs and `test:read-only` goes back to covering just the tutorial suite.

`npm test` and `test:all` now run all three files, and the PR job in test-tutorials.yml calls `npm test` so the new guards run on pull requests.
